Move code into ui-components
Moves code that is used by both code snippets and data sources extension for inserting code into the ui-components package.
